Start organizing local vs cloud
