+++ pyrisco/common.py
@@ -1,0 +1,12 @@
+GROUP_ID_TO_NAME = ["A", "B", "C", "D"]
+
+class UnauthorizedError(Exception):
+    """Exception to indicate an error in authorization."""
+
+
+class CannotConnectError(Exception):
+    """Exception to indicate an error in authorization."""
+
+
+class OperationError(Exception):
+    """Exception to indicate an error in operation."""
